add react-router-dom + fix main.scss
