+++ package.json
@@ -21,6 +21,7 @@
       "react-dnd-html5-backend": "^16.0.1",
       "react-dom": "^18.2.0",
       "react-redux": "^8.0.5",
+      "react-router-dom": "^6.8.0",
       "react-scripts": "5.0.1",
       "redux": "^4.2.0",
       "typescript": "^4.9.4",
