+++ src/components/App.jsx
@@ -1,42 +1,30 @@
 import AppHeader from "./AppHeader/AppHeader";
-import BurgerIngredients from "./BurgerIngredients/BurgerIngredients";
-import BurgerConstructor from "./BurgerConstructor/BurgerConstructor";
-import { DndProvider } from "react-dnd";
-import { HTML5Backend } from "react-dnd-html5-backend";
-import { useEffect } from "react";
-import { getIngredients } from "../services/slices/burgerIngredientsSlice";
-import { useDispatch, useSelector } from "react-redux";
+import { useSelector } from "react-redux";
 import Loader from "./Loader/Loader";
+import { BrowserRouter, Routes, Route } from "react-router-dom";
+import HomePage from "../pages/HomePage";
+import ErrorPage from "../pages/404";
 
 
 function App() {
-   const dispatch = useDispatch()
-   const loaderVisible = useSelector( store => store.loader.visible)
-
-   // получаем данные ингредиентов
-   useEffect( () => {
-      dispatch( getIngredients() )
-   }, [] )
+   const loaderVisible = useSelector( store => store.loader.visible )
 
 
    return (
-      <div>
+      <>
          { loaderVisible && <Loader/> }
 
          <AppHeader/>
          <main className="main wrapper pt-10 pb-10">
-            <h1 className="text text_type_main-large mb-5">Соберите бургер</h1>
-
-            <div className="main-row">
-               <DndProvider backend={ HTML5Backend }>
-                  <BurgerIngredients/>
-                  <BurgerConstructor/>
-               </DndProvider>
-            </div>
-
+            <BrowserRouter>
+               <Routes>
+                  <Route path={ '/' } element={ <HomePage/> }/>
+                  <Route path={ '*' } element={ <ErrorPage/> }/>
+               </Routes>
+            </BrowserRouter>
          </main>
 
-      </div>
+      </>
    )
 }
