+++ src/pages/404.jsx
@@ -1,0 +1,12 @@
+import React from 'react'
+
+
+function ErrorPage()
+{
+   return (
+      <div>404</div>
+   )
+}
+
+
+export default ErrorPage
